requirements.txt
customtkinter
Pillow
numpy
opencv-python
matplotlib